+++ R/catalogs.R
@@ -10,7 +10,8 @@
 #' @return A `tibble`.
 #'
 #'
-#' @source [Time-series bulk data
+#' @source
+#' [Time-series bulk data
 #' download](https://www.bde.es/webbe/en/estadisticas/recursos/descargas-completas.html)
 #'
 #' @param catalog A single value indicating the catalogs to be updated
@@ -208,7 +209,8 @@ bde_catalog_load <-
 #'
 #' @return None. Downloads the catalog file(s) to the local machine.
 #'
-#' @source [Time-series bulk data
+#' @source
+#' [Time-series bulk data
 #' download](https://www.bde.es/webbe/en/estadisticas/recursos/descargas-completas.html)
 #'
 #' @param catalog A vector of characters indicating the catalogs to be updated