+++ R/scales.R
@@ -1,25 +1,24 @@
-#' BdE scales for `ggplot2`.
+#' BdE scales for \CRANpkg{ggplot2}.
 #'
 #' @description
 #'
 #' Scales to be used with the \CRANpkg{ggplot2} package. Discrete palettes are
 #' named as `scale_*_bde_d` while continuous palettes are named `scale_*_bde_c`.
 #'
-#' @seealso [bde_vivid_pal()], [ggplot2::discrete_scale()],
-#'   [ggplot2::continuous_scale()]
+#' @seealso [ggplot2::discrete_scale()], [ggplot2::continuous_scale()]
 #'
 #' @family bde_plot
 #'
 #' @export
 #'
-#' @return A `ggplot2` color scale.
+#' @return A \CRANpkg{ggplot2} color scale.
 #'
 #' @rdname scales_bde
 #'
 #' @name scales_bde
 #'
 #' @param palette Name of the BdE palette to apply. One of `"bde_vivid_pal"`,
-#'   `"bde_rose_pal"`. See [bde_pals()] for details.
+#'   `"bde_rose_pal"`. See [bde_palettes()] for details.
 #'
 #' @param ... Further arguments of [ggplot2::discrete_scale()] or
 #'   [ggplot2::continuous_scale()].
@@ -48,69 +47,56 @@
 #'   theme_minimal()
 #'
 scale_color_bde_d <- function(palette = c("bde_vivid_pal", "bde_rose_pal"),
-                              ...) { # nocov start
-
+                              ...) {
   palette <- match.arg(palette)
-  valid_pals <- c("bde_vivid_pal", "bde_rose_pal")
 
-  if (!palette %in% valid_pals) {
-    stop("palette should be one of: ", paste0(valid_pals, collapse = ", "))
-  }
+  cols_v <- bde_palettes(palette = palette)
+  pal <- scales::manual_pal(cols_v)
 
-  cols <- switch(palette,
-    "bde_vivid_pal" = bde_vivid_pal(),
-    "bde_rose_pal" = bde_rose_pal()
-  )
   ggplot2::discrete_scale(
     aesthetics = "color",
+    palette = pal,
     scale_name = palette,
-    palette = cols,
     ...
   )
-  # nocov end
 }
+
+#' @rdname scales_bde
+#' @name scales_bde
+#' @export
+#' @usage NULL
+scale_colour_bde_d <- scale_color_bde_d
 
 #' @rdname scales_bde
 #' @name scales_bde
 #' @export
 scale_fill_bde_d <- function(palette = c("bde_vivid_pal", "bde_rose_pal"),
-                             ...) { # nocov start
+                             ...) {
   palette <- match.arg(palette)
-  valid_pals <- c("bde_vivid_pal", "bde_rose_pal")
 
-  if (!palette %in% valid_pals) {
-    stop("palette should be one of :", paste0(valid_pals, collapse = ", "))
-  }
+  cols_v <- bde_palettes(palette = palette)
+  pal <- scales::manual_pal(cols_v)
 
-  cols <- switch(palette,
-    "bde_vivid_pal" = bde_vivid_pal(),
-    "bde_rose_pal" = bde_rose_pal()
-  )
   ggplot2::discrete_scale(
     aesthetics = "fill",
+    palette = pal,
     scale_name = palette,
-    palette = cols,
     ...
   )
-  # nocov end
 }
 
 
 #' @rdname scales_bde
 #' @name scales_bde
 #' @export
-scale_color_bde_c <- function(palette = c("bde_vivid_pal", "bde_rose_pal"),
-                              ...) { # nocov start
+scale_color_bde_c <- function(palette = c("bde_rose_pal", "bde_vivid_pal"),
+                              ...) {
   palette <- match.arg(palette)
-  valid_pals <- c("bde_vivid_pal", "bde_rose_pal")
 
-  if (!palette %in% valid_pals) {
-    stop("palette should be one of :", paste0(valid_pals, collapse = ", "))
-  }
 
   cols <- switch(palette,
-    "bde_vivid_pal" = bde_vivid_pal()(6),
-    "bde_rose_pal" = c(bde_rose_pal()(6)[1:3], bde_rose_pal()(6)[6:4])
+    "bde_vivid_pal" = bde_palettes(6, "bde_vivid_pal"),
+    "bde_rose_pal" = bde_palettes(6, "bde_rose_pal")[c(1, 2, 3, 6, 5, 4)]
   )
   ggplot2::continuous_scale(
     aesthetics = "color",
@@ -118,24 +104,25 @@ scale_color_bde_c <- function(palette = c("bde_vivid_pal", "bde_rose_pal"),
     palette = scales::gradient_n_pal(cols),
     ...
   )
-  # nocov end
 }
 
 #' @rdname scales_bde
 #' @name scales_bde
 #' @export
-scale_fill_bde_c <- function(palette = c("bde_vivid_pal", "bde_rose_pal"),
-                             ...) { # nocov start
-  palette <- match.arg(palette)
-  valid_pals <- c("bde_vivid_pal", "bde_rose_pal")
+#' @usage NULL
+scale_colour_bde_c <- scale_color_bde_c
 
-  if (!palette %in% valid_pals) {
-    stop("palette should be one of :", paste0(valid_pals, collapse = ", "))
-  }
+#' @rdname scales_bde
+#' @name scales_bde
+#' @export
+scale_fill_bde_c <- function(palette = c("bde_rose_pal", "bde_vivid_pal"),
+                             ...) {
+  palette <- match.arg(palette)
+
 
   cols <- switch(palette,
-    "bde_vivid_pal" = bde_vivid_pal()(6),
-    "bde_rose_pal" = c(bde_rose_pal()(6)[1:3], bde_rose_pal()(6)[6:4])
+    "bde_vivid_pal" = bde_palettes(6, "bde_vivid_pal"),
+    "bde_rose_pal" = bde_palettes(6, "bde_rose_pal")[c(1, 2, 3, 6, 5, 4)]
   )
   ggplot2::continuous_scale(
     aesthetics = "fill",
@@ -143,5 +130,4 @@ scale_fill_bde_c <- function(palette = c("bde_vivid_pal", "bde_rose_pal"),
     palette = scales::gradient_n_pal(cols),
     ...
   )
-  # nocov end
 }