+++ R/tidyBdE-package.R
@@ -4,10 +4,12 @@
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
+#' @importFrom lifecycle deprecated
 ## usethis namespace: end
 NULL
 
 # import stuffs
 #' @importFrom utils download.file read.csv2
+#' @importFrom tibble tibble
 #' @importFrom ggplot2 unit element_rect element_text element_blank element_line
 NULL